1. 几乎完成代码了，就差上报拼数据的部分了

+++ src/clock.h
@@ -10,13 +10,13 @@ typedef struct _clock_struct_t
   uint32_t hour;   // 0 - 23
   uint8_t min;    // 0 - 59
   uint8_t sec;    // 0 - 59
-  uint8_t ms625;   // 0 - 511
+  uint16_t ms625;
 } clock_struct_t;
 
 uint32_t clock_get_hour(void);
 uint8_t clock_get_min(void);
 uint8_t clock_get_sec(void);
-uint8_t clock_get_ms625(void);
+uint16_t clock_get_ms625(void);
 
 void clock_initialize(void);
 void clock_time_proc(enum task_events ev);
